vmtests: ensure tmpdir is cleaned between test class changes
vmtests refactoring exposed an issue with the amount of storage
required to complete a full make vmtest.  At it's peak, it requires
48G of storage space due to the base class __del__ function not
being invoked when switching between testcase classes in a test
file.

This patch fixes this by moving the delete logic into a method
on the TempDir class and adding a call to the new method in to
base class' tearDownClass classmethod.

bzr-revno: 291.2.1

+++ tests/vmtests/__init__.py
@@ -133,10 +133,14 @@ class TempDir:
         subprocess.check_call(['tar', '-C', self.mnt, '-xf', self.output_disk],
                               stdout=DEVNULL, stderr=subprocess.STDOUT)
 
-    def __del__(self):
+    def remove_tmpdir(self):
         if (os.getenv('KEEP_VMTEST_DATA', "false") != "true"):
             # remove tempdir
-            shutil.rmtree(self.tmpdir)
+            if os.path.exists(self.tmpdir):
+                shutil.rmtree(self.tmpdir)
+
+    def __del__(self):
+        self.remove_tmpdir()
 
 
 class VMBaseClass:
@@ -283,7 +287,8 @@ class VMBaseClass:
 
     @classmethod
     def tearDownClass(self):
-        logger.debug('Removing launch logfile')
+        logger.debug('Removing class tmpdir: {}'.format(self.td.tmpdir))
+        self.td.remove_tmpdir()
 
     @classmethod
     def expected_interfaces(self):
